agent-service/requirements.txt
langgraph==0.4.8
langchain-openai==0.3.18
langchain-core==0.3.51
simpleeval==1.0.3
grpcio==1.78.0
grpcio-tools==1.78.0
grpcio-health-checking==1.78.0
protobuf==6.31.1
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-instrumentation-grpc==0.61b0
opentelemetry-instrumentation-httpx==0.61b0
